python/requirements.txt
GitPython>=3.1.0
requests>=2.31.0
